Add files via upload
The goal of a hand-written digit classification project is to develop a machine learning model that can accurately classify images of hand-written digits. This is a popular problem in the field of computer vision, and is often used as a benchmark task for evaluating the performance of machine learning algorithms.